used long long int for the flop counts instead of long int

+++ src/GPU/Matrices/dbl_baqr_kernels.cu
@@ -67,8 +67,8 @@ __global__ void dbl_small_house
 }
 
 void flopcount_dbl_small_house
- ( int dim, int dimLog2, long int *add, long int *mul, long int *div,
-   long int *sqrtfun )
+ ( int dim, int dimLog2, long long int *add, long long int *mul, long long int *div,
+   long long int *sqrtfun )
 {
    *add += dimLog2 + 4;
    *mul += dim + 3;
@@ -155,8 +155,8 @@ __global__ void cmplx_small_house
 }
 
 void flopcount_cmplx_small_house
- ( int dim, int dimLog2, long int *add, long int *mul, long int *div,
-   long int *sqrtfun )
+ ( int dim, int dimLog2, long long int *add, long long int *mul, long long int *div,
+   long long int *sqrtfun )
 {
    *add += 3*dim + dimLog2 + 6;
    *mul += 6*dim + 6;
@@ -197,7 +197,7 @@ __global__ void dbl_small_leftRupdate
 }
 
 void flopcount_dbl_small_leftRupdate
- ( int nrows, int ncols, int szt, int k, long int *add, long int *mul )
+ ( int nrows, int ncols, int szt, int k, long long int *add, long long int *mul )
 {
    const int nbthreads = ncols - k;
 
@@ -255,7 +255,7 @@ __global__ void cmplx_small_leftRupdate
 }
 
 void flopcount_cmplx_small_leftRupdate
- ( int nrows, int ncols, int szt, int k, long int *add, long int *mul )
+ ( int nrows, int ncols, int szt, int k, long long int *add, long long int *mul )
 {
    const int nbthreads = ncols - k;
 
@@ -288,7 +288,7 @@ __global__ void dbl_small_betaRTv
 }
 
 void flopcount_dbl_small_betaRTv 
- ( int nrows, int ncols, int szt, int k, long int *add, long int *mul )
+ ( int nrows, int ncols, int szt, int k, long long int *add, long long int *mul )
 {
    const int nbthreads = nrows - k;
 
@@ -333,7 +333,7 @@ __global__ void cmplx_small_betaRHv
 }
 
 void flopcount_cmplx_small_betaRHv 
- ( int nrows, int ncols, int szt, int k, long int *add, long int *mul )
+ ( int nrows, int ncols, int szt, int k, long long int *add, long long int *mul )
 {
    const int nbthreads = nrows - k;
 
@@ -404,7 +404,7 @@ __global__ void dbl_medium_subvbetaRTv
 }
 
 void flopcount_dbl_medium_subvbetaRTv
- ( int nrows, int ncols, int szt, int k, long int *add, long int *mul )
+ ( int nrows, int ncols, int szt, int k, long long int *add, long long int *mul )
 {
    const int endcol = (k+1)*szt;     // 1 + last column index in tile
    // total number of entries in R that will be modified
@@ -458,7 +458,7 @@ __global__ void cmplx_medium_subvbetaRHv
 }
 
 void flopcount_cmplx_medium_subvbetaRHv
- ( int nrows, int ncols, int szt, int k, long int *add, long int *mul )
+ ( int nrows, int ncols, int szt, int k, long long int *add, long long int *mul )
 {
    const int endcol = (k+1)*szt;     // 1 + last column index in tile
    // total number of entries in R that will be modified
@@ -513,7 +513,7 @@ __global__ void dbl_VB_to_W
 }
 
 void flopcount_dbl_VB_to_W
- ( int nrows, int ncols, long int *add, long int *mul )
+ ( int nrows, int ncols, long long int *add, long long int *mul )
 {
    // the number of threads equals nrows
    *add += nrows*((ncols-1)*ncols/2*(1+nrows) + ncols-1);
@@ -598,7 +598,7 @@ __global__ void cmplx_VB_to_W
 }
 
 void flopcount_cmplx_VB_to_W
- ( int nrows, int ncols, long int *add, long int *mul )
+ ( int nrows, int ncols, long long int *add, long long int *mul )
 {
    // the number of threads equals nrows
    *add += nrows*((ncols-1)*ncols/2*(6+2*nrows) + 2*(ncols-1));
@@ -622,7 +622,7 @@ __global__ void dbl_beta_times_V
    if(idx < nrows) W[idx] = result;
 }
 
-void flopcount_dbl_beta_times_V ( int nrows, long int *mul )
+void flopcount_dbl_beta_times_V ( int nrows, long long int *mul )
 {
    const int nbthreads = nrows;
    // equals rowdim as in the call to the dbl_beta_times kernel
@@ -655,7 +655,7 @@ __global__ void cmplx_beta_times_V
    }
 }
 
-void flopcount_cmplx_beta_times_V ( int nrows, long int *mul )
+void flopcount_cmplx_beta_times_V ( int nrows, long long int *mul )
 {
    const int nbthreads = nrows;
    // equals rowdim as in the call to the dbl_beta_times kernel
@@ -679,7 +679,7 @@ __global__ void dbl_initialize_WYT
    if(idx < dim*dim) WYT[idx] = result;
 }
 
-void flopcount_dbl_initialize_WYT ( int dim, long int *mul )
+void flopcount_dbl_initialize_WYT ( int dim, long long int *mul )
 {
    // the number of threads equals dim*dim
    *mul += dim*dim;
@@ -711,7 +711,7 @@ __global__ void cmplx_initialize_WYH
 }
 
 void flopcount_cmplx_initialize_WYH
- ( int dim, long int *add, long int *mul )
+ ( int dim, long long int *add, long long int *mul )
 {
    // the number of threads equals dim*dim
    *add += 2*dim*dim;
@@ -737,7 +737,7 @@ __global__ void dbl_update_WYT
 }
 
 void flopcount_dbl_update_WYT
- ( int dim, long int *add, long int *mul )
+ ( int dim, long long int *add, long long int *mul )
 {
    // the number of threads equals dim*dim
    *add += dim*dim;
@@ -774,7 +774,7 @@ __global__ void cmplx_update_WYH
 }
 
 void flopcount_cmplx_update_WYH
- ( int dim, long int *add, long int *mul )
+ ( int dim, long long int *add, long long int *mul )
 {
    // the number of threads equals dim*dim
    *add += 4*dim*dim;
@@ -831,7 +831,7 @@ __global__ void dbl_beta_next_W
    if(idx < nrows) W[idx] = result;
 }
 
-void flopcount_dbl_beta_next_W ( int nrows, long int *add, long int *mul )
+void flopcount_dbl_beta_next_W ( int nrows, long long int *add, long long int *mul )
 {
    // the number of threads equals nrows
    *add += nrows;
@@ -909,7 +909,7 @@ __global__ void cmplx_beta_next_W
    }
 }
 
-void flopcount_cmplx_beta_next_W ( int nrows, long int *add, long int *mul )
+void flopcount_cmplx_beta_next_W ( int nrows, long long int *add, long long int *mul )
 {
    // the number of threads equals nrows
    *add += 4*nrows;
@@ -940,7 +940,7 @@ __global__ void dbl_small_WYT
 }
 
 void flopcount_dbl_small_WYT
- ( int nrows, int szt, long int *add, long int *mul )
+ ( int nrows, int szt, long long int *add, long long int *mul )
 {
    const int nbthreads = nrows*nrows;
 
@@ -981,7 +981,7 @@ __global__ void cmplx_small_WYH
 }
 
 void flopcount_cmplx_small_WYH
- ( int nrows, int szt, long int *add, long int *mul )
+ ( int nrows, int szt, long long int *add, long long int *mul )
 {
    const int nbthreads = nrows*nrows;
 
@@ -1013,7 +1013,7 @@ __global__ void dbl_small_QWYT
 }
 
 void flopcount_dbl_small_QWYT
- ( int dim, int rowdim, int szt, int coloff, long int *add, long int *mul )
+ ( int dim, int rowdim, int szt, int coloff, long long int *add, long long int *mul )
 {
    const int nbthreads = dim*rowdim;
 
@@ -1055,7 +1055,7 @@ __global__ void cmplx_small_QWYH
 }
 
 void flopcount_cmplx_small_QWYH
- ( int dim, int rowdim, int szt, int coloff, long int *add, long int *mul )
+ ( int dim, int rowdim, int szt, int coloff, long long int *add, long long int *mul )
 {
    const int nbthreads = dim*rowdim;
 
@@ -1088,7 +1088,7 @@ __global__ void dbl_small_YWTC
 }
 
 void flopcount_dbl_small_YWTC
- ( int rowdim, int coldim, long int *add, long int *mul )
+ ( int rowdim, int coldim, long long int *add, long long int *mul )
 {
    const int nbthreads = rowdim*coldim;
 
@@ -1131,7 +1131,7 @@ __global__ void cmplx_small_YWHC
 }
 
 void flopcount_cmplx_small_YWHC
- ( int rowdim, int coldim, long int *add, long int *mul )
+ ( int rowdim, int coldim, long long int *add, long long int *mul )
 {
    const int nbthreads = rowdim*coldim;
 
@@ -1160,7 +1160,7 @@ __global__ void dbl_small_Qupdate
 }
 
 void flopcount_dbl_small_Qupdate
- ( int dim, int rowdim, long int *add )
+ ( int dim, int rowdim, long long int *add )
 {
    const int nbthreads = dim*rowdim;
 
@@ -1193,7 +1193,7 @@ __global__ void cmplx_small_Qupdate
 }
 
 void flopcount_cmplx_small_Qupdate
- ( int dim, int rowdim, long int *add )
+ ( int dim, int rowdim, long long int *add )
 {
    const int nbthreads = dim*rowdim;
 
@@ -1222,7 +1222,7 @@ __global__ void dbl_small_R_add_YWTC
 }
 
 void flopcount_dbl_small_R_add_YWTC
- ( int nrows, int coldim, int szt, int rowoff, int coloff, long int *add )
+ ( int nrows, int coldim, int szt, int rowoff, int coloff, long long int *add )
 {
    const int rowdim = nrows - rowoff;
    const int nbthreads = rowdim*coldim;
@@ -1256,7 +1256,7 @@ __global__ void cmplx_small_R_add_YWHC
 }
 
 void flopcount_cmplx_small_R_add_YWHC
- ( int nrows, int coldim, int szt, int rowoff, int coloff, long int *add )
+ ( int nrows, int coldim, int szt, int rowoff, int coloff, long long int *add )
 {
    const int rowdim = nrows - rowoff;
    const int nbthreads = rowdim*coldim;
@@ -1269,8 +1269,8 @@ void GPU_dbl_small_house
    int colidx, int nrows1, int k, int L,
    double *A_h, double *A_d,
    double *v_h, double *V_d, double *beta_h, double *beta_d,
-   double *lapms, long int *add, long int *mul, long int *div,
-   long int *sqrtfun, bool verbose )
+   double *lapms, long long int *add, long long int *mul, long long int *div,
+   long long int *sqrtfun, bool verbose )
 {
    const int nrLog2 = ceil(log2((double) nrows1));
    const int rowidx = colidx*(nrows+1);       // start of number in A_h
@@ -1337,8 +1337,8 @@ void GPU_cmplx_small_house
    double *Are_h, double *Aim_h, double *Are_d, double *Aim_d,
    double *vre_h, double *vim_h, double *Vre_d, double *Vim_d,
    double *beta_h, double *beta_d,
-   double *lapms, long int *add, long int *mul, long int *div,
-   long int *sqrtfun, bool verbose )
+   double *lapms, long long int *add, long long int *mul, long long int *div,
+   long long int *sqrtfun, bool verbose )
 {
    const int nrLog2 = ceil(log2((double) nrows1));
    const int rowidx = colidx*(nrows+1);       // start of number in A_h
@@ -1413,7 +1413,7 @@ void GPU_cmplx_small_house
 void GPU_dbl_small_leftRupdate
  ( int nrows, int ncols, int szt, int colidx, int k, int L,
    double *A_h, double *A_d, double *V_d, double *beta_h, double *beta_d,
-   double *lapms, long int *add, long int *mul, long int *div,
+   double *lapms, long long int *add, long long int *mul, long long int *div,
    bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
@@ -1453,7 +1453,7 @@ void GPU_cmplx_small_leftRupdate
  ( int nrows, int ncols, int szt, int colidx, int k, int L,
    double *Are_h, double *Aim_h, double *Are_d, double *Aim_d,
    double *Vre_d, double *Vim_d, double *beta_h, double *beta_d,
-   double *lapms, long int *add, long int *mul, long int *div,
+   double *lapms, long long int *add, long long int *mul, long long int *div,
    bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
@@ -1493,7 +1493,7 @@ void GPU_dbl_medium_leftRupdate
  ( int nrows, int ncols, int szt, int colidx, int k, int L,
    double *A_h, double *A_d, double *V_d, double *beta_h, double *beta_d,
    double *w_h, double *w_d,
-   double *lapms, long int *add, long int *mul, long int *div,
+   double *lapms, long long int *add, long long int *mul, long long int *div,
    bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
@@ -1556,7 +1556,7 @@ void GPU_cmplx_medium_leftRupdate
    double *Are_h, double *Aim_h, double *Are_d, double *Aim_d,
    double *Vre_d, double *Vim_d, double *beta_h, double *beta_d,
    double *wre_h, double *wim_h, double *wre_d, double *wim_d,
-   double *lapms, long int *add, long int *mul, long int *div,
+   double *lapms, long long int *add, long long int *mul, long long int *div,
    bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
@@ -1625,7 +1625,7 @@ void GPU_dbl_VB_to_W
  ( int nrows, int ncols, int szt,
    double *V_h, double *V_d, double *W_h, double *W_d,
    double *beta_h, double *beta_d, double *lapms,
-   long int *add, long int *mul, long int *div, bool verbose )
+   long long int *add, long long int *mul, long long int *div, bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
    cudaEventCreate(&start);
@@ -1671,7 +1671,7 @@ void GPU_cmplx_VB_to_W
    double *Vre_h, double *Vim_h, double *Vre_d, double *Vim_d,
    double *Wre_h, double *Wim_h, double *Wre_d, double *Wim_d,
    double *beta_h, double *beta_d, double *lapms,
-   long int *add, long int *mul, long int *div, bool verbose )
+   long long int *add, long long int *mul, long long int *div, bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
    cudaEventCreate(&start);
@@ -1744,7 +1744,7 @@ void GPU_dbl_medium_VB_to_W
  ( int nrows, int ncols, int szt, int idx,
    double *V_h, double *V_d, double *W_h, double *W_d,
    double *WYT_h, double *WYT_d, double *beta_h, double *beta_d,
-   double *lapms, long int *add, long int *mul, long int *div,
+   double *lapms, long long int *add, long long int *mul, long long int *div,
    bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
@@ -1835,7 +1835,7 @@ void GPU_cmplx_medium_VB_to_W
    double *Wre_h, double *Wim_h, double *Wre_d, double *Wim_d,
    double *WYHre_h, double *WYHim_h, double *WYHre_d, double *WYHim_d,
    double *beta_h, double *beta_d,
-   double *lapms, long int *add, long int *mul, long int *div,
+   double *lapms, long long int *add, long long int *mul, long long int *div,
    bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
@@ -1958,7 +1958,7 @@ void GPU_cmplx_medium_VB_to_W
 
 void GPU_dbl_small_WYT
  ( int nrows, int szt, double *W_d, double *Y_d, double *WYT_d,
-   double *WYT_h, double *lapms, long int *add, long int *mul, long int *div,
+   double *WYT_h, double *lapms, long long int *add, long long int *mul, long long int *div,
    bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
@@ -1994,7 +1994,7 @@ void GPU_cmplx_small_WYH
  ( int nrows, int szt, double *Wre_d, double *Wim_d,
    double *Yre_d, double *Yim_d, double *WYHre_d, double *WYHim_d,
    double *WYHre_h, double *WYHim_h, double *lapms,
-   long int *add, long int *mul, long int *div, bool verbose )
+   long long int *add, long long int *mul, long long int *div, bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
    cudaEventCreate(&start);
@@ -2032,7 +2032,7 @@ void GPU_cmplx_small_WYH
 
 void GPU_dbl_small_YWT
  ( int nrows, int szt, int idx, double *Y_d, double *W_d, double *YWT_d,
-   double *YWT_h, double *lapms, long int *add, long int *mul, long int *div,
+   double *YWT_h, double *lapms, long long int *add, long long int *mul, long long int *div,
    bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
@@ -2069,7 +2069,7 @@ void GPU_cmplx_small_YWH
  ( int nrows, int szt, int idx,
    double *Yre_d, double *Yim_d, double *Wre_d, double *Wim_d,
    double *YWHre_d, double *YWHim_d, double *YWHre_h, double *YWHim_h,
-   double *lapms, long int *add, long int *mul, long int *div, bool verbose )
+   double *lapms, long long int *add, long long int *mul, long long int *div, bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
    cudaEventCreate(&start);
@@ -2109,7 +2109,7 @@ void GPU_cmplx_small_YWH
 void GPU_dbl_small_QWYT
  ( int dim, int szt, int idx, double *Q_d, double *WYT_d, double *QWYT_d,
    double *QWYT_h, double *Q_h, double *lapms,
-   long int *add, long int *mul, long int *div, bool verbose )
+   long long int *add, long long int *mul, long long int *div, bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
    cudaEventCreate(&start);
@@ -2160,7 +2160,7 @@ void GPU_cmplx_small_QWYH
  ( int dim, int szt, int idx, double *Qre_d, double *Qim_d,
    double *WYHre_d, double *WYHim_d, double *QWYHre_d, double *QWYHim_d,
    double *QWYHre_h, double *QWYHim_h, double *Qre_h, double *Qim_h,
-   double *lapms, long int *add, long int *mul, long int *div, bool verbose )
+   double *lapms, long long int *add, long long int *mul, long long int *div, bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
    cudaEventCreate(&start);
@@ -2220,7 +2220,7 @@ void GPU_cmplx_small_QWYH
 void GPU_dbl_small_YWTC
  ( int nrows, int ncols, int szt, int idx,
    double *YWT_d, double *C_d, double *YWTC_d, double *YWTC_h,
-   double *lapms, long int *add, long int *mul, long int *div,
+   double *lapms, long long int *add, long long int *mul, long long int *div,
    bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
@@ -2287,7 +2287,7 @@ void GPU_cmplx_small_YWHC
  ( int nrows, int ncols, int szt, int idx,
    double *YWHre_d, double *YWHim_d, double *Cre_d, double *Cim_d,
    double *YWHCre_d, double *YWHCim_d, double *YWHCre_h, double *YWHCim_h,
-   double *lapms, long int *add, long int *mul, long int *div, bool verbose )
+   double *lapms, long long int *add, long long int *mul, long long int *div, bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
    cudaEventCreate(&start);
@@ -2356,7 +2356,7 @@ void GPU_cmplx_small_YWHC
 
 void GPU_dbl_small_Qupdate
  ( int dim, int szt, int idx, double *Q_d, double *QWYT_d,
-   double *Q_h, double *lapms, long int *add, long int *mul, long int *div,
+   double *Q_h, double *lapms, long long int *add, long long int *mul, long long int *div,
    bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
@@ -2393,7 +2393,7 @@ void GPU_dbl_small_Qupdate
 void GPU_cmplx_small_Qupdate
  ( int dim, int szt, int idx, double *Qre_d, double *Qim_d,
    double *QWYHre_d, double *QWYHim_d, double *Qre_h, double *Qim_h,
-   double *lapms, long int *add, long int *mul, long int *div, bool verbose )
+   double *lapms, long long int *add, long long int *mul, long long int *div, bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
    cudaEventCreate(&start);
@@ -2433,7 +2433,7 @@ void GPU_cmplx_small_Qupdate
 
 void GPU_dbl_small_R_add_YWTC
  ( int nrows, int ncols, int szt, int idx, double *R_d, double *YWTC_d,
-   double *R_h, double *lapms, long int *add, long int *mul, long int *div,
+   double *R_h, double *lapms, long long int *add, long long int *mul, long long int *div,
    bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
@@ -2473,7 +2473,7 @@ void GPU_cmplx_small_R_add_YWHC
  ( int nrows, int ncols, int szt, int idx,
    double *Rre_d, double *Rim_d, double *YWHCre_d, double *YWHCim_d,
    double *Rre_h, double *Rim_h, double *lapms,
-   long int *add, long int *mul, long int *div, bool verbose )
+   long long int *add, long long int *mul, long long int *div, bool verbose )
 {
    cudaEvent_t start,stop;           // to measure time spent by kernels 
    cudaEventCreate(&start);
@@ -2517,8 +2517,8 @@ void GPU_dbl_blocked_houseqr
    double *houselapms, double *tileRlapms, double *vb2Wlapms,
    double *WYTlapms, double *QWYTlapms, double *Qaddlapms,
    double *YWTlapms, double *YWTClapms, double *Raddlapms,
-   double *walltimesec, long int *addcnt, long int *mulcnt,
-   long int *divcnt, long int *sqrtcnt, bool verbose )
+   double *walltimesec, long long int *addcnt, long long int *mulcnt,
+   long long int *divcnt, long long int *sqrtcnt, bool verbose )
 {
    const int dim = nrows*ncols;         // total number of doubles
    const int nrows2 = nrows*nrows;
@@ -2684,8 +2684,8 @@ void GPU_cmplx_blocked_houseqr
    double *houselapms, double *tileRlapms, double *vb2Wlapms,
    double *WYTlapms, double *QWYTlapms, double *Qaddlapms,
    double *YWTlapms, double *YWTClapms, double *Raddlapms,
-   double *walltimesec, long int *addcnt, long int *mulcnt,
-   long int *divcnt, long int *sqrtcnt, bool verbose )
+   double *walltimesec, long long int *addcnt, long long int *mulcnt,
+   long long int *divcnt, long long int *sqrtcnt, bool verbose )
 {
    const int dim = nrows*ncols;           // total number of doubles
    const int nrows2 = nrows*nrows;
